run ktlint and detekt for main files

+++ app/src/main/java/com/cradle/neptune/view/SplashActivity.kt
@@ -10,14 +10,17 @@ class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_splash)
-        Handler().postDelayed({
-            val intent = Intent(this@SplashActivity, LoginActivity::class.java)
-            startActivity(intent)
-            finish()
-        }, DELAY.toLong())
+        Handler().postDelayed(
+            {
+                val intent = Intent(this@SplashActivity, LoginActivity::class.java)
+                startActivity(intent)
+                finish()
+            },
+            DELAY
+        )
     }
 
     companion object {
-        private const val DELAY = 100
+        private const val DELAY = 100L
     }
 }
